Trying new stuff with lavaLampExperiment. Didn't like work out
I feel like I can't use the useClickHoverWander hook too much.  It gets boring because it's the same method of interacting with something.  After a few things based on this, you lose the sense of play with trying to figure out how something works and how you can interact with it.  It's already known.